move message handling to folder and refactor
